+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2023-05-04
+
+### Added
+
+- Regularization to the Networking (Dropout, L1 and L2)
+- Newtonsoft.Json dll to allow Architecture gathering
+
 ## [0.8.0] - 2023-05-03
 
 ### Changed
@@ -122,7 +129,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PointComputations.cs, to facilitate convex hull computation, point cloud simplification, and continuing on, any solely point-based methods.
 - Conversion.cs, to allow conversion of Vector types.
 
-[unreleased]: https://github.com/IrishFix/Computational-Geometry/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/IrishFix/Computational-Geometry/compare/v1.0.0...HEAD
+[0.8.0]: https://github.com/IrishFix/Computational-Geometry/compare/v0.8.0...v1.0.0
 [0.8.0]: https://github.com/IrishFix/Computational-Geometry/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/IrishFix/Computational-Geometry/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/IrishFix/Computational-Geometry/compare/v0.5.0...v0.6.0